python-api/requirements.txt
fastapi[all]
uvicorn[standard]
httpx
pyjwt
passlib[bcrypt]
python-multipart
pydantic-settings
